\name{centrarSalida}
\alias{centrarSalida}
\title{ Force plot coordinates }
\description{
  For internal use of POIPlot function
}
\usage{
centrarSalida()
}

\details{
  For internal use of poiPLOT. Force plot coordinates to {0,0}
}

\author{ 
  Eduardo San Miguel Martin 
} 

\seealso{\code{\link{fishIin}}, \code{\link{fishIout}}, \code{\link{POIPlot-methods}}}
