\name{circulo}
\alias{circulo}
\title{ Function to plot a circle }
\description{
  Internal function. Not intended to be called directly by user.
}
\usage{
circulo(cx, cy, r, circleCol, PLOT = TRUE)
}

\arguments{
  \item{cx}{x coordinate for circle center }
  \item{cy}{ y coordinate for circle center }
  \item{r}{ Radius }
  \item{circleCol}{ Circle color to be plotted }
  \item{PLOT}{ Should circle be plotted? }
}
\details{
    Not to be called directly by user.
}
\value{
   A matrix containing circle coordinates to be plotted.
}

\author{ 
  Eduardo San Miguel Martin 
} 

\seealso{\code{\link{circulin}}, \code{\link{POIPlot-methods}}}
