\name{POIcalculate<--methods}
\docType{methods}
\alias{POIcalculate<--methods}
\title{ Methods for Function POIcalculate<-  }
\description{
 Methods for function \code{POIcalculate<-} 
}
\section{Methods}{
\describe{

\item{object = "multiPOI"}{ Replacement method }

\item{object = "POI"}{ Replacement method }
}}

