\name{POIcalculate<-}
\alias{POIcalculate<-}

\title{ Replacement function for POI(Pcoords,PcoordsFI,newPcoords,objeto) slots }
\description{
   Replacement function for POI(Pcoords,PcoordsFI,newPcoords,objeto) slots
}
\usage{
POIcalculate(object) <- value
}

\arguments{
  \item{object}{ Object of class POI }
  \item{value}{ Values to fill slots with }
}
\details{
  To be used with POICalc
}
\value{
   POI object with updated slots
}
\seealso{\code{\link{POIPlot-methods}},\code{\link{POI-class}},\code{\link{plotPOI}}}

