\name{POIcolors<-}
\alias{POIcolors<-}

\title{ POI-class slot Colores Replacement function }
\description{
  Fill colores slot with appropiate value.
}
\usage{
POIcolors(object) <- value
}
\arguments{
  \item{object}{ Object of class POI }
  \item{value}{ Object of class Vector with colors to apply }
}
\value{
   Replace slot colores in object
}
