\name{POIcoords<--methods}
\docType{methods}
\alias{POIcoords<--methods}

\title{ Methods for Function POIcoords<-  }
\description{
  Methods for function \code{POIcoords<-}  
}
\section{Methods}{
\describe{
\item{object = "POI"}{ Method that allow POI to update its Pcoords, PcoordsFI, newPcoords and objeto Slots }
}}
\keyword{methods}
