\name{POIPlot-methods}
\docType{methods}
\alias{POIPlot-methods}
\title{ Methods for Function POIPlot}
\description{
 Methods for function \code{POIPlot}  
}
\section{Methods}{
\describe{
\item{POI = "POI"}{ See function documentation for details. }

\item{POI = "multiPOI"}{ See function documentation for details. }

\item{POI = "POIGraph"}{ See function documentation for details. }
}}
\keyword{methods}

